Apply SWC transforms for Middleware also to 3rd-party code
This aligns Turbopack with what Webpack is already doing in this regard.

What this doesn't enable though is emitting build errors when using the
disallowed APIs. Just enabling it for 3rd-party code in the SWC
transforms wouldn't be sufficient. This would lead to too many false
positives, because it's applied before tree shaking. We would need to
change that first so that the detection and emitting errors is done
after tree-shaking, as is already the case with Webpack.