Backport: Fix `unstable_allowDynamic` when used with pnpm
When using dependencies in Middleware that make use of dynamic code
evaluation, Next.js emits a build error because this is not supported in
the Edge runtime.

In rare cases, when the code can not be reached at runtime and can't be
removed by tree-shaking, users might opt in to using the
`unstable_allowDynamic` config.

When combined with pnpm, the provided glob patterns as documented at
https://nextjs.org/docs/messages/edge-dynamic-code-evaluation#possible-ways-to-fix-it
did not match correctly because of pnpm's use of the `.pnpm` directory.

To fix the pattern matching, we need to provide the `dot` option to
[picomatch](https://github.com/micromatch/picomatch), which enables
dotfile matching.

_Side note: Ideally we would detect dynamic code evaluation after tree
shaking, to reduce the number of cases where users need to revert to
using `unstable_allowDynamic`._

fixes #51401