[Segment Cache] No data on tree prefetch if no PPR
When the client Segment Cache attempts to prefetch a route tree, and it
the route does not have PPR enabled, the server falls through to the old
implementation, which sends back a FlightRouterState.

This is fine, since the client can use the FlightRouterState to recreate
the non-PPR behavior.

However, when loading.tsx is present in the route, the server will not
just send back the FlightRouterState, but also render the page up to the
first loading boundary. This is not what we want in the Segment Cache
implementation — we want the route tree only, so we can determine which
parts the client already has before sending a second request for
the rest.

So this PR disables the loading.tsx mechanism when prefetching the
route tree, by checking the Next-Router-Segment-Prefetch header.