Increase test retries from 1 to 2 (#73795)
Let's see how this affects the failure rate and duration of our tests in
CI.

Test times for any given group might increase a bit, while hopefully at
least increasing the chance that the whole test group succeeds. This
would in theory reduce the number of times we have to re-run a whole
failed test group, which is quite time consuming.

A possible downside of this change is that flaky tests which have a
chance of getting fixed with a better implementation might be harder to
detect.