[Segment Cache] Add PPR header to segment prefetch (#73756)
Follow-up to #73715. Since PPR should always result in a cache hit for a
prefetch (even in the worst case, there's a fallback cache entry), the
only reason a route tree prefetch would not be found in the per-segment
cache is because PPR is disabled.

So we should just fallthrough to the old prefetching flow whenever PPR
is disabled.

But the client still needs some way to tell whether the response was
served from the per-segment cache or by the old implementation.

So I added an explicit response header to every per-segment prefetch. If
the header is missing, the client can infer that the route does not
support PPR.