Retry manifest file loading only in dev mode
In #73891 we added another manifest to be loaded in `loadComponents`.
This uncovered a flakiness in prod mode when attempting to load an
optional manifest. The non-existent manifest is attempted to be loaded
three times with 100ms delay between attempts, before giving up. For
some reason this increased loading time leads to more test flakiness.

To mitigate this, we're limiting the retry behaviour to the dev mode,
which matches the original intention when this was introduced in #45244.