Remove parentRendered argument (#73877)
It doesn't look like this argument is logically needed. No callers pass
it as an option except for the recursive call, and the recursive call
will never set it to `true` because of the `renderComponentsOnThisLevel`
branch that happens earlier in the function.

I'm guessing it was needed at one point but then the logic changed.