Allow usage of `this` and `arguments` in nested function expression
In #73059 we added build-time checks for the forbidden usage of `this`
and `arguments` in server functions. A nested function expression is
however allowed to use these expressions.

x-ref: #74124

To add a bit more context: `this` is forbidden in server actions because
the Next.js compiler hoists inline server actions out of their original
location into the module scope, so that they can be imported and invoked
by the action handler. Due to this hoisting, the `this` context gets
lost and is not available to the server action.

To prevent surprising runtime errors for such cases, we emit a build
error to provide feedback to developers as early as possible.

However, nested function declarations or function expressions create
a new `this` context, and in those it's allowed to access `this` and
`arguments`.