used shared worker for lint & typecheck steps (#74154)
Similar to #73138, this removes the direct usage of `jest-worker` for
the linting/typechecking step in favor of the shared worker that has
built-in handling for propagating errors to the parent process. This is
to ensure that if the worker performing the typechecking/linting
receives a SIGKILL signal (which could happen with an OOM error), that
the parent process exits appropriately.