[Turbopack] fix import.meta.url in monorepo (#72612)
### What?

fixes the computation of `import.meta.url` inside of monorepos (when the root_path != project_path)

### How?

We add a new argument to the ChunkingContext that represents the relative path from the output folder to the root of the monorepo. Since they are on different filesystems, we can't compute them from the values alone.