Add `value` to PrefetchSubtaskResult
PrefetchSubtaskResult is a type used to spawn an non-blocking async task
while also providing a promise that represents when the task is
complete, so that the scheduler can limit the number of concurrent
network requests.

This updates the type to also include a generic `value` field, so the
task can return additional data to the caller. This must be separate
from the `closed` field so that the caller can unwrap it before the
entire operation has completed. For example, the `value` field will
usually represent a streaming result (like a Flight promise) that
resolves  immediately, whereas the `closed` field only resolves when the
stream is complete.

No behavior changes in this commit. I'll use this to implement cache
revalidation in the following steps.