feat: connect error rating buttons to telemetry API (#74496)
We're adding a new feature to the error overlay that allows developers
to rate the current error message. Developers can rate the error message
by clicking on the thumb up or down icon in the error overlay footer.

This pull request adds the click handler for the error rating buttons.
When the buttons are clicked, the click handler calls a previously
implemented internal API endpoint. The endpoint is connected to the
usual Next.js Telemetry.

To test the UI, run `pnpm storybook` and go to
http://localhost:6006/?path=/story/erroroverlaylayout--default.



https://github.com/user-attachments/assets/6daf6e64-e8bb-4e5e-b282-082086f1e529