Indicate boolean value for configured experimental features on startup
When printing the configured experimental features of the Next.js
config, we are currently not discriminating `true` from `false` values.
This is especially confusing when disabling an experimental feature that
is enabled by default. In this case it appears in the output as if the
feature was enabled.

By using the `✓` and `⨯` for boolean feature flags (and `·` for others),
users can now clearly see whether a configured feature is enabled or
disabled.