[Segment Cache] Fix stale time unit conversion
Fixes a bug in the handling of stale times in the client Segment Cache,
when prefetching a PPR-enabled route.

The stale time sent by the server is in seconds, but I was handling it
as if it were in milliseconds. This caused the entry to expire
almost immediately.

I didn't notice this sooner because the main test apps I've been working
with doesn't have PPR enabled, and the non-PPR path was converting the
values correctly. The existing e2e tests also didn't catch it because
they run fast enough that they the entries don't expire.