feat: added rewrite headers after user-supplied rewrites (#74776)
Next.js's client router needs to know the pathname and query that the
request was rewritten to in order to facilitate reuse of static RSC
payloads generated from fallbacks. This takes the form of additional
headers being sent back on the response that includes the correct
rewritten pathname that later the client can take into account when
generating the client route key.